# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Enactus Platform API
  version: 0.4.0
  description: |
    API de Enactus Platform (eduXaction Colombia).

    **Formato de error único.** Todo error responde con la misma forma:

    ```json
    { "error": { "code": "forbidden", "message": "…", "details": … } }
    ```

    `code` es el identificador estable que el cliente mira para decidir qué
    hacer; `message` es para la persona. Un error inesperado se registra
    entero en el servidor y llega genérico al cliente — nunca un detalle
    interno de PostgreSQL.

    **Dos criterios que conviene tener presentes al leer:**

    - Pedir por id un recurso que existe pero está fuera del alcance del rol
      devuelve **404, no 403**. Un 403 ya confirmaría que ese recurso existe.
    - Un estudiante de Open Learning que llama un endpoint exclusivo de
      Enactus recibe **403, nunca una lista vacía**: una lista vacía haría
      ver un bug de permisos como si fuera "todavía no hay datos".

    **Alcance de lectura.** Ningún listado devuelve todo a todo el mundo: el
    filtro por rol se aplica en el servidor y no se puede cambiar con un
    parámetro de la petición.
servers:
  - url: http://localhost:3000
    description: Desarrollo local
  - url: https://api-staging.eduxaction.co
    description: Staging (pendiente, Fase 6)
  - url: https://api.eduxaction.co
    description: Producción (pendiente, Fase 6)
...
security:
  - bearerAuth: []
paths:
...
  /students/{id}/ruta-progress:
    parameters: [{ $ref: '#/components/parameters/id' }]
    get:
      tags: [students]
      summary: Ruta de Impacto completa
      description: |
        Laboratorios, fases, módulos, objetivos y cursos, con el desbloqueo y
        el estado de cada deadline.

        Un estudiante de Open Learning recibe **403**. Si quien pregunta es
        del equipo pero el estudiante objetivo es de Open Learning, responde
        **409** diciéndolo — nunca una lista vacía.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RutaProgress' }
        '403': { description: Cuenta de Open Learning, o no es su propio progreso }
        '409': { description: El estudiante objetivo es de Open Learning }
...
  /certificates:
    get:
      tags: [certificates]
      summary: Certificados
      parameters:
        - { $ref: '#/components/parameters/page' }
        - { $ref: '#/components/parameters/pageSize' }
        - { in: query, name: studentId, schema: { type: string, format: uuid } }
      responses: { '200': { description: Página de certificados } }
  /certificates/ruta:
    post:
      tags: [certificates]
      summary: Emitir el certificado de Ruta de Impacto
      description: |
        Doble candado a propósito:

        1. El endpoint valida que las 3 fases estén completas y, si no,
           responde **409 con el detalle de qué falta**.
        2. La base lo hace cumplir igual con el trigger
           `certificates_require_complete_ruta`.

        Requiere `can_grade_enactus` — la Ruta es contexto Enactus, no alcanza
        con el permiso de Open Learning. Admin y superadmin lo hacen sin pasar
        por el permiso, y queda en `audit_log`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentId, laboratoryId]
              properties:
                studentId: { type: string, format: uuid }
                laboratoryId: { type: string, format: uuid }
      responses:
        '201': { description: Emitido }
        '403': { description: Sin permiso de calificar en Enactus }
        '409':
          description: Ruta incompleta, o el certificado ya existía
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /admin/users/{id}/can-grade:
    parameters: [{ $ref: '#/components/parameters/id' }]
    patch:
      tags: [admin]
      summary: Cambiar el permiso de calificar de un LXD
      description: |
        Son **dos** permisos, no uno. Cada cambio queda en `audit_log` con
        quién, a quién, cuándo, y los valores anterior y nuevo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                canGradeOpenLearning: { type: boolean }
                canGradeEnactus: { type: boolean }
      responses:
        '200': { description: Usuario actualizado }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: El permiso solo aplica a cuentas LXD }
...
  /admin/restore:
    post:
      tags: [admin]
      summary: Restaurar un respaldo
      description: |
        El endpoint más peligroso de la API. **Solo superadmin**, transaccional
        (o entra todo, o no entra nada), y exige la confirmación literal
        `REEMPLAZAR TODOS LOS DATOS`. Queda en `audit_log`.

        Efecto secundario deliberado: al reemplazar `users` se caen las
        sesiones abiertas de todo el mundo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version, data, confirm]
              properties:
                version: { type: integer }
                data: { type: object }
                confirm: { type: string, enum: ['REEMPLAZAR TODOS LOS DATOS'] }
      responses:
        '200': { description: Filas restauradas por tabla }
        '400': { description: Falta la confirmación exacta }
        '403': { description: Solo superadmin }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token de 12 h. Se renueva con `/auth/refresh`.
  parameters:
    id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      in: query
      name: pageSize
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
...
    Forbidden:
      description: El rol no tiene permiso
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - bad_request
                - validation_error
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - payload_too_large
                - too_many_requests
                - storage_not_configured
                - internal_error
            message: { type: string }
            details: {}
...
    RutaProgress:
      type: object
      properties:
        studentId: { type: string, format: uuid }
        studentName: { type: string }
        laboratories:
          type: array
          items:
            type: object
            properties:
              laboratoryId: { type: string, format: uuid }
              laboratoryName: { type: string }
              phasesTotal: { type: integer }
              phasesDone: { type: integer }
              isComplete: { type: boolean }
              certificateIssued: { type: boolean }
              phases:
                type: array
                items:
                  type: object
                  properties:
                    phaseId: { type: string, format: uuid }
                    orderIndex: { type: integer }
                    title: { type: string }
                    deadline: { type: string, format: date, nullable: true }
                    deadlineStatus:
                      type: string
                      enum: [none, on_track, approaching, overdue]
                      description: Una fase completada nunca aparece atrasada.
                    modulesTotal: { type: integer }
                    modulesDone: { type: integer }
                    isComplete: { type: boolean }
                    isUnlocked: { type: boolean }
                    objectives: { type: array, items: {} }
                    modules: { type: array, items: {} }